ref(InviteModal): Refactor InviteModalHook into its own file
