+++ static/app/components/modals/memberInviteModalCustomization.tsx
@@ -1,0 +1,11 @@
+import HookOrDefault from 'sentry/components/hookOrDefault';
+
+export const InviteModalHook = HookOrDefault({
+  hookName: 'member-invite-modal:customization',
+  defaultComponent: ({onSendInvites, children}) =>
+    children({sendInvites: onSendInvites, canSend: true}),
+});
+
+export type InviteModalRenderFunc = React.ComponentProps<
+  typeof InviteModalHook
+>['children'];
