+++ static/app/views/settings/organizationMembers/inviteRequestRow.tsx
@@ -3,7 +3,10 @@ import styled from '@emotion/styled';
 
 import {Button} from 'sentry/components/button';
 import Confirm from 'sentry/components/confirm';
-import HookOrDefault from 'sentry/components/hookOrDefault';
+import {
+  InviteModalHook,
+  InviteModalRenderFunc,
+} from 'sentry/components/modals/memberInviteModalCustomization';
 import PanelItem from 'sentry/components/panels/panelItem';
 import RoleSelectControl from 'sentry/components/roleSelectControl';
 import Tag from 'sentry/components/tag';
@@ -23,14 +26,6 @@ type Props = {
   onUpdate: (data: Partial<Member>) => void;
   organization: Organization;
 };
-
-const InviteModalHook = HookOrDefault({
-  hookName: 'member-invite-modal:customization',
-  defaultComponent: ({onSendInvites, children}) =>
-    children({sendInvites: onSendInvites, canSend: true}),
-});
-
-type InviteModalRenderFunc = React.ComponentProps<typeof InviteModalHook>['children'];
 
 function InviteRequestRow({
   inviteRequest,
